moved IPartyDetail interface to shared
