+++ crash/src/shared/iparty-details.ts
@@ -1,0 +1,8 @@
+export interface IPartyDetails {
+    FirstName: string;
+    LastName: string;
+    Address: string;
+    Phone: string;
+    License: string;
+    Remarks: string;
+  }
